requirements.txt
flask==3.1.2
requests-html==0.10.0
pymupdf==1.26.7
python-dotenv==1.2.1
gunicorn==25.1.0
Werkzeug==3.1.5
pdf2docx==0.5.8
pytesseract==0.3.13
Pillow==11.2.1
python-docx==1.1.2